requirements.txt
jmcomic
pyzipper
Pillow
fpdf
